auto-submit/requirements.txt
# Azure Functions Core
azure-functions>=1.18.0

# Azure Cosmos DB SDK
azure-cosmos>=4.5.1

# Azure Identity for Managed Identity authentication
azure-identity>=1.15.0

# HTTP client for making requests
aiohttp>=3.9.0

# Additional utilities
python-dateutil>=2.8.2

# Note: azure-functions-worker is provided by Azure Functions runtime
# Only include it if you need specific versions for local testing
